matchbox-keyboard-ui.c: Fix int -> long to address 32bit vs 64 bit addressing
The soft keyboard was displaying incorrectly on 64bit machines

[YOCTO #9254]

+++ src/matchbox-keyboard-ui.c
@@ -125,7 +125,7 @@ get_desktop_area(MBKeyboardUI *ui, int *x, int *y, int *width, int *height)
   Atom           atom_area, type;
   int            result, format;
   unsigned long  nitems, bytes_after;
-  int           *geometry = NULL;
+  long           *geometry = NULL;
 
   atom_area = XInternAtom (ui->xdpy, "_NET_WORKAREA", False);
